update price in ShoppingItem.kt class From double to Float because float is 32 bits but double is 64 bits and not necessary in here.

+++ app/src/main/java/com/example/chopchopapplication/data/local/ShoppingItem.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "shopping_table")
 data class ShoppingItem(
     var name: String,
-    var price: Double,
+    var price: Float,
     var count: Int
 ) {
     @PrimaryKey(autoGenerate = true) var id: Int? = null
